% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogram.R
\name{geom_hist}
\alias{geom_hist}
\title{ogram}
\usage{
geom_hist(dataset, variable)
}
\description{
This function changes the geom_histogram() function in ggplot2 to the new geom_hist() function.
}
\details{
Ogram: How to be Lazy
}
